Moved Tuple3 into util package
